Refacto SendModal to be able to block/allow close it with keyboard/click
